pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "faith-cli"
version = "0.1.0"
description = "Framework AI Team Hive"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "AGPL-3.0" }
authors = [
  { name = "FAITH Contributors" }
]
dependencies = [
  "click>=8.1",
  "fastapi>=0.110,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic>=2.0,<3.0",
  "python-dotenv>=1.0,<2.0",
  "pyyaml>=6.0,<7.0",
  "redis>=5.0,<6.0",
  "requests>=2.31",
  "uvicorn[standard]>=0.29,<1.0",
  "websocket-client>=1.6",
]

[project.optional-dependencies]
dev = [
  "jinja2>=3.1",
  "pre-commit>=4.0",
  "pytest>=9.0",
  "pytest-asyncio>=1.0",
  "ruff>=0.11",
]

[project.scripts]
faith = "faith_cli.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["faith*", "faith_cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
  ".git",
  ".pytest_cache",
  "faith_cli.egg-info",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
